+++ x64BareBones/Userland/shell/lib/syscall_call.c
@@ -5,7 +5,7 @@ extern int sys_write(int fb, const char* buf);
 extern int sys_read (char *buf);
 extern void sys_clearwindow(uint32_t* color);
 extern int sys_date_time (int tpye);
-extern double sys_benchmark (int type);
+extern uint64_t  sys_benchmark (int type);
 
 
 #define STDIN   0
@@ -44,13 +44,13 @@ void get_date(){
 }
 
 
-double do_benchmark_fps(){
+uint64_t  do_benchmark_fps(){
     sys_benchmark (0);
 }
-double do_benchmark_floating_point(){
+uint64_t  do_benchmark_floating_point(){
     sys_benchmark (1);
 }
-double do_benchmark_hardware_access(){
+uint64_t  do_benchmark_hardware_access(){
     sys_benchmark (2);
 }
